requirement.txt
requests
numpy
Pillow
pma_python
multiresolutionimageinterface
PyYAML
